fix `git git st` when not at top-level dir of repo
0d8e5a7773afbcbfbd54be4c66b9789a32ee99c9 has a bug: "If the alias
expansion is prefixed with an exclamation point, it will be treated as a
shell command. [...] Note that shell commands will be executed from the
top-level directory of a repository, which may not necessarily be the
current directory. GIT_PREFIX is set as returned by running git
rev-parse --show-prefix from the original current directory."
(`get help config`, then search for 'alias')

Now we get to the fun part: my first attempt was to just do

	git = !cd "$GIT_PREFIX" && git

which worked great when NOT at top-level dir of repo, but if you WERE at
the top-level, it would complain about not being in a git repo--
apparently, it was cd-ing to my homedir, because at the top-level,
$GIT_PREFIX == '', and so it was doing `cd   && git st`. Note that at
the Bash prompt, `cd ""` in fact stays where you are and does not cd to
your homedir, and `cd "$GIT_PREFIX_OR_ANY_UNDEFINED_VAR" && git st`
works fine, so I was a little confused. This did suggest to me the
clever solution[1] of

	git = !echo $(cd "GIT_PREFIX") > /dev/null && git

Luckily I kept googling and found the solution that's actually getting
committed.

[1]: I actually first tried

	git = !echo $(cd "$GIT_PREFIX") && git

which worked great but printed out an extra newline before everything,
so then naturally I tried

	git = !echo -n $(cd "$GIT_PREFIX") && git

which, to my surprise, printed '-n' on its own line before everything.

Now, git parsing double-quoted variable substitutions in the shell
commands, rather than letting Bash handle it; not entirely sure why
they'd do that, but they probably had their reasons, this is a pretty
obscure edge case, whatever. But what the fuck did they do to echo?