Improve CtrlP config for big repos
Building the cache is slow, don't clear it on exit, I'll manually tell
you to.

Increase max number of results from 10 to 40, because it seems that
CtrlP will stop searching as soon as it reaches 10 results, even if
there's a much better result that it hasn't seen yet (like, say, an
exact file name match >:( ). I tried 0 at first based on
  https://github.com/ctrlpvim/ctrlp.vim/issues/332
but it was *incredibly* sluggish, probably because when typing the first
few letters there are tons and tons of results; even at 100 it felt a
little laggy. 40 doesn't quite feel snappy but at least it's usually not
noticeably slow.